fix/MAF-6 refactor

+++ src/main/java/com/example/mafgan/domain/user/UserService.java
@@ -1,12 +1,14 @@
 package com.example.mafgan.domain.user;
 
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
-@Service
+@Service // TODO ZROB METODY W KOLEJNOSCI CRUD
 public class UserService {
     private final UserRepository userRepository;
 
@@ -14,39 +16,31 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User findById(final Long id) {
-        return userRepository.findById(id).orElse(null);
+    public Optional<User> findById(final Long id) {
+        return userRepository.findById(id);
     }
 
-    public User add(final User user) {
+    public User add(@NonNull final User user) { // TODO SAVE
+        user.setIdUser(null);
+//        if (user.getIdUser() != null){ //TODO Piotrek niech ocenni co jest zgrabniejsze
+//            throw new RuntimeException("User to save should not have user id");
+//        } else {
+        return userRepository.save(user); }
+
+    public User update(final User user) { //TODO SPRAWDZ CZY JEST USER najlepiej bez ifa
         return userRepository.save(user);
     }
 
-    public User update(
-            User toUpdate
-    ) {
-        return userRepository.save(toUpdate);
-    }
-
     public void deleteById(final Long id) {
-        userRepository.deleteById(id);
-    }
-
-    public void delete(User user) {
-        userRepository.delete(user);
+        if (userRepository.existsById(id)){
+            userRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("User not exist");
+        }
     }
 
     public boolean existsById(final Long id) {
         return userRepository.existsById(id);
     }
 
-    public void addRoleToUser(Long id, String role) {
-        User user = userRepository.findById(id).get();
-        user.addRole(UserRole.valueOf(role));
-    }
-
-    public void deleteRoleInUser(Long id, String role) {
-        User user = userRepository.findById(id).get();
-        user.deleteRole(UserRole.valueOf(role));
-    }
 }
